# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booklab - backend",
        "description": "Booklab project API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "operationId": "f92ff81e405f6f6224ed353490975ced",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "musk@email.com",
                                    "password": "password1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Logout",
                "operationId": "ac1c8a886ecd17e108b62c43687439ed",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Register",
                "operationId": "e7e369b43858339ddf6d345e5dfdf7a1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Musk",
                                    "email": "musk@email.com",
                                    "password": "password1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/cardapios": {
            "post": {
                "tags": [
                    "Cardapios"
                ],
                "summary": "Add new Cardapio",
                "operationId": "61c6eaf598f5fb97bf9557b8021184ab",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "foto": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "decimal"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "titulo": "Almoço",
                                    "descricao": "Prato Feito",
                                    "foto": "imagem.jpg",
                                    "preco": 1.89,
                                    "categoria_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cardapios/{id}?with[]=categoria&with[]=items": {
            "get": {
                "tags": [
                    "Cardapios"
                ],
                "summary": "Mostra Cardápio e items do cardápio",
                "description": "Show Cardápio",
                "operationId": "ab20029f25c3d3c440c9824a0d2b75e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/cardapios/{id}": {
            "put": {
                "tags": [
                    "Cardapios"
                ],
                "summary": "Updates a Cardapio",
                "operationId": "2f3befdb38897a563b5c1f34ba9a8c02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update Cardapio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "estado_id",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "foto": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "decimal"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "titulo": "Almoço",
                                    "descricao": "Prato Feito",
                                    "foto": "imagem.jpg",
                                    "preco": 1.89,
                                    "categoria_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cardapios"
                ],
                "summary": "Delete Cardapio",
                "description": "Delete Cardapio",
                "operationId": "214b9550e21f5164d13a37f728b3471a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}